chore : GH-30 Add People feature
