+++ .editorconfig
@@ -1,4 +1,5 @@
 [*.{kt,kts}]
-ktlint_code_style = ktlint_official
+ktlint_code_style = android_studio
 insert_final_newline = true
+no-line-break-before-assignment = disabled
 ktlint_function_naming_ignore_when_annotated_with = Composable
