+++ .idea/codeStyles/Project.xml
@@ -5,6 +5,7 @@
     </JetCodeStyleSettings>
     <codeStyleSettings language="kotlin">
       <option name="CODE_STYLE_DEFAULTS" value="KOTLIN_OFFICIAL" />
+      <option name="ASSIGNMENT_WRAP" value="5" />
     </codeStyleSettings>
   </code_scheme>
 </component>
