+++ .idea/gradle.xml
@@ -38,6 +38,7 @@
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/movieDetails" />
             <option value="$PROJECT_DIR$/feature/movies" />
+            <option value="$PROJECT_DIR$/feature/people" />
             <option value="$PROJECT_DIR$/lint" />
             <option value="$PROJECT_DIR$/ui-test-hilt-manifest" />
           </set>
