+++ feature/movieDetails/src/test/java/me/jerryokafor/feature/moviedetails/screen/MoviesDetailsScreenTest.kt
@@ -41,12 +41,12 @@ import me.jerryokafor.feature.moviedetails.viewmodel.TEST_ID
 import me.jerryokafor.ihenkiri.core.test.rule.assertAreDisplayed
 import me.jerryokafor.ihenkiri.core.test.util.FakeImageLoader
 import me.jerryokafor.ihenkiri.core.test.util.MovieDetailsTestData
-import me.jerryokafor.ihenkiri.feature.moviedetails.MOVIE_DETAILS_BOTTOM_BAR
-import me.jerryokafor.ihenkiri.feature.moviedetails.MOVIE_DETAILS_COL
-import me.jerryokafor.ihenkiri.feature.moviedetails.MOVIE_DETAILS_MAIN_CAST
-import me.jerryokafor.ihenkiri.feature.moviedetails.MOVIE_DETAILS_OVERVIEW
-import me.jerryokafor.ihenkiri.feature.moviedetails.MoviesDetailViewModel
-import me.jerryokafor.ihenkiri.feature.moviedetails.MoviesDetails
+import me.jerryokafor.ihenkiri.feature.moviedetails.ui.MOVIE_DETAILS_BOTTOM_BAR
+import me.jerryokafor.ihenkiri.feature.moviedetails.ui.MOVIE_DETAILS_COL
+import me.jerryokafor.ihenkiri.feature.moviedetails.ui.MOVIE_DETAILS_MAIN_CAST
+import me.jerryokafor.ihenkiri.feature.moviedetails.ui.MOVIE_DETAILS_OVERVIEW
+import me.jerryokafor.ihenkiri.feature.moviedetails.ui.MoviesDetails
+import me.jerryokafor.ihenkiri.feature.moviedetails.viewmodel.MoviesDetailViewModel
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
